old/requirements.txt
psycopg2==2.7
django-councilmatic<0.8.1
